Added UI when we have no Hub, Device or Scene. (#19)
Added UI to guide the user when we have not found the Hub or have no
Device or Scene. Suggest next step for the user to add a device,
discover devices, or create a scene.
To support the UI, I simplified initialization of `Items` in
`ItemListViewModel`'s and added a UI bindable `IsEmpty` property.
I renamed "Click" to "Tap" in several places to be a bit more mobile
friendly.
I added notifications when the device picture changes to populate the
list of devices progressively while discovering devices.