Only record changes if needed. (#27)
This change clarifies the ability to start/stop model change recording.
Recording now starts off, so that house models on which
`House.StartRecordingChanges` is not called will not record. This
prevents from doing unnecessary work for models used transiently, such
as when saving the house configuration for example.
Renamed `House.PlayModel` to `PlayChanges` for clarity.