Make "Sync" and "Save" buttons in status bar optional (#28)
With this change we make the "Sync" and "Save" buttons on the bottom
right of the window an optional feature, designed mostly for debug.
These buttons introduce a user confirmation for syncing the house
configuration with network devices or saving the house configuration to
a local or online file. They are intended to help the developer confirm
the state before syncing or saving when working on features under
development.

These buttons can be shown and enabled by adding the following to
`appsettings.json`:
```
  "FeatureFlags": {
    "EnableHouseSaveButton": true,
    "EnableHouseSyncButton": true,
  },
```

Details:
- Now obtain whether to show "Sync" and "Save" button in status bar
using `appsettings.json`.
- Hide and show button appropriately and directly trigger the "Sync" or
"Save" on model changes if not shown.
- Add some debouncing logic to avoid saving too often.
- Improved `UpdatePropertiesSyncStatus` (in Device and Channel) to
distinguish between model change and device sync and defer calling it
when setting multiple properties in a row.
- Removed unnecessary properties and methods from `StatusBar.xaml.cs`
and use the `SettingsViewModel` directly instead.
- Fixed bugs in implementation of `TryWriteRampRateAsync`.