Fix: device would not show connected after pressing the Connect button (#30)
When a device showed "not connected" and the connect button was pressed,
the device would be connected but the status would not be refreshed on
screen. This is because the `isKnownConnectionStatus` is set to true,
preventing `ScheduleGetConnectionStatus` from acquiring the new status.

This change sets the connection status directly when adding or
connecting a device, since we already know at this point that the device
is connected and therefore do not have to rely on
`ScheduleGetConnectionStatus` in that case.

Since we can also set the `isKnownConnectionStatus` flag in that case,
in addition to the status, I added a small helper function to set both
callable from all the places where we set a known connection status.