Fixed crash when adding a new gateway device (#32)
Adding a new hub sends a DeviceInserted notification to the device view
and this would trip an exception in all cases. In fact, we can handle
that notification for a hub since there is no view change for adding a
new hub. We also enforce that the hub device is added at the top of the
list.