Fixed issues that prevented the app to work on WASM (#34)
Some issues were preventing the app to work on WASM
- The redirect URL for OneDrive authentication was incorrect
- On WASM, the `SelectTemplateCore` method `container` parameter passes
in the `ContentPresenter`, not its container as on Windows and Android.
- That method can also be called with `TemplatedParent` on the
`ContentPresenter`. Sometimes `Parent` is not null and can be used,
sometimes both are null and all we can do is return with no template
(null). We seem to be called again in that case. Note that this seems to
be specific to WASM. These situations don't occur with WindowsAppSdk or
on Android as far as I can tell.