Fixed overactive assert and a bug in a conditional. (#47)
In Insteon.Device.UpdatePropertiesSyncStatus :
Fixed overactive assert 
Fixed logic to ensure that that we set the sync status properly when we
don't have a device driver or have not read the properties from the
physical device, and when we have properties, but the values differ
between model and physical device.
Refer to the comments in the code for details.