Fixed a few build warnings (#48)
Fixed a couple nullable build warnings, 
Attempted to fix
"desktop-targetframework-must-be-placed-first-in-the-targetframeworks-property"
without success but left the code and reference to the doc in place in
case this starts to work with future versions of Uno,
Removed now unused msbuild code that used to patch the build result to
avoid duplication of a WebView2 asset at deployment time. This appears
to have gotten fixed by WinUI team.