Fixed a couple issues targeting "Desktop" on Windows (#49)
Fixed a crash where the binding code was trying to dereference `NavView`
before it was loaded. Added `AppShell.CompactModeThresholdWidth` to
handle the null case. No exactly sure why this happened only for Desktop
but this must have to do with the way this specific binding works.

Also disabled expanding window content to overlap the title bar to have
a custom title bar. This is not working on Desktop and yield incorrect
rendering.