Fixed issues while adding new devices (#74)
This change fixes regressions that in some cases prevented adding new
devices, and in other cases prevented the device list item to update
when the user was changing its properties.
- Fixed object type comparison in `DeviceViewModel.GetOrCreate`
- Added missing logic to `AddDeviceBtnClick` to properly cover the case
where a device view model already exists but is of the wrong type,
usually because we created a vanilla `DeviceViewModel` earlier (before
we knew the type of the view model) to help increase UI responsiveness.
We need to remove it from the presented list, and add the new one.
- Removed creation of a device view model in `DeviceListViewModel`, as
this would duplicate the one created in `IDevicesObserver.DeviceAdded`