Ensured no new device was created when cancelling out of New Device dialog. (#77)
If the device was auto-discovered, it would be added to the list of
devices before "Add" or "Cancel" were pressed in the dialog but not removed if the
user cancelled out of the dialog. All that was necessary was to propagate
the flag indicating this is a new device or not when auto-discovering. Fixed.