Defer querying new device properties to improve perceived performance (#78)
Fixes #70 closes #70 
Instead of querying device and channel properties immediately before
completing the add device operation, we schedule these requests for
later and complete the add device operation earlier. This helps the
perceived performance in the New Device dialog when auto-discovering a
new device or when pressing the primary button.