Allow binding to all OperatingFlags properties, including binding KeypadButtonView to number of buttons (#79)
After PR #78, the `KeypadButtonView` of the new device would sometimes
have the wrong number of buttons because `OperatingFlags`, which stores
the bit property indicating 6 or 8 buttons (Is8Button), was not yet read
from the device when the button view was first created, and dynamically
binding to Is8Button later was not implemented,

This change alters the XAML in `KeypadView` to be able to switch between
6 and 8 buttons views dynamically based on the Is8Button property. It
also makes it possible to bind to any other property stored in
`OperatingFlags` and to get property change notifications to fire for
the affected properties when `OperatingFlags` changes.

To simplify the code and make it less error proof, this change
introduces a static array containing the description of each property
stored as a bit in `OperatingFlags`, i.e., the name of that property and
its index in `OperatingFlags` (or `OpFlags2`).

This change also ensures that the current selected button in the
`KeypadButtonView` stays coherent when switching from 8 to 6 buttons or
vice versa.