Improving following command response. (#80)
This change improves the likeliness of success of parsing the response
of a command by handling two specific cases:
- Sometimes in a command expecting a standard response, an extended
message is sent first by the device. We used to stop parsing the
response and eventually time out. We now ignore the extended message and
move forward in the hope the standard response will follow. It appears
that it usually does.
- When discovering a device with the `StartIMAllLinking` command, and
either the user is instructed to press and hold the SET button on the
device or the device Id is communicated to the IM via an
`EnterLinkingMode` command, other devices appear to sometimes send
standard messages to the IM and the IM to echo these messages. Though we
handled the messages being sent, we used to stall and eventually time
out on the echoes. Now we just ignore the echo and move forward.
Eventually, we get to the response from the device we are trying to
all-link.