Improved design for Add, Remove, Replace, Copy devices and tested them (#81)
This change improves the design for Add, Remove, Replace, and Copy
Device. These operations need to throw some UI before execution. This is
now done in the App/UI layer by binding the UI to the code behind the
pages originating these operations, then control is passed to the View
Model layer to complete the work. This simplifies the flow with no need
to hop back from the View Model to the App/UI.

In this change:
- Gathered Add, Remove, Replace, Copy Device functionality in
`DeviceViewModel`. Since these operations are independent of the
presented list view of devices (if any), I removed them from
`DeviceListViewModel`.

- Now report to observers of the list of devices that the product data
(type) of a device has changed. For example, a generic device can be
created as placeholder for perceived performance and then changed to be
a keypad. `DeviceListViewModel` listens to this notification and can
recreate the device view model and swap it in the list.

- Removed access to the New Device Dialog from the View Model layer.
These are now only access from the App/UI (see above).

- Tested that these operations work as expected, and fixed a few issues.