Cleanup and small enhancements determining device Connection Status. (#90)
A set of name changes, cleanups and small enhancements:
- Added "Async" to `TryReadchannelProperties` and
`TryWriteChannelProperties` for naming consistency
- Renamed `IsProductDataRead` to `IsProductDataKnown` since this is set
by deserializing as well as by reading product data from the device
- Made `force` attribute work on `GetConnectionStatus` and
`ScheduleGetConnectionStatus`. This allows to re-ping the device even if
we had pinged before, which is now the case when navigating to the
device view.
- Introduced Device level methods for reading/writing the properties of
a given channel to be able to check on connection status, and for
consistency with similar methods.
- Simplified `ScheduleSync` by utilizing `DeviceNeedsSync` to bail out
immediately if no sync is needed.
- Conditioned Read/Write methods on device being "reachable", i.e.,
connected and hub operational.
- Clarified some commentary