Added ability to run a prehandler for all jobs. (#91)
The Scheduler logs a starting message regardless of whether the job will
do anything. Many jobs will fail immediately if the device is not
connected, without attempting to actually run the job. Yet, a starting
message is logged at information level and shows in the status bar,
giving the user the impression that something happened.

This change adds the ability to have pre-handler called before calling
the job handler. The pre-handler can return false to indicate major
failure and skip running the job. This will result in no
information-level logging for the job, so nothing will show up in the
status bar or info-level console.