Request to re-acquire connectivity status at the end of each read/write to a device (#94)
After attempting to read or write from/to a device, request to
re-acquire the connectivity status of the device if the read or write
operation failed. This will force a ping of the device to acquire the
connectivity status on the next read or write operation and prevent
future read/write operations with that device until the device is
reconnected.