More efficient handling of device Connection Status (#97)
We now have the following flow for the Connection Status of a device:

`RetrieveConnectionStatusAsync` (renamed form
`GetConnectionStatusAsync`) retrieves the connection status by pinging
the device if the connection status is not believed to be up to date. It
sets the result in `Device.Status` and also sets
`isConnectionStatusKnown` to indicate Status is believed to be up to
date.

This method is called before attempting any other read/write operation
with the device. If the result is not `Connected` the operation and
subsequent ones to the same device are prevented, saving useless
attempts at interacting with the device.

Once the problem is believed to be fixed, the user should go to the view
for the device and reconnect the device.

Any time a read or write operation fails, `isConnectionStatusKnown` is
reset to force a new retrieval. If that determines that the device is
indeed not connected, no subsequent operations to the same device will
be attempted.

Also renamed `ScheduleGetConnectionStatus` to
`ScheduleRetrieveConnectionStatus` for consistency.