Ensured that Gateways and Scenes where properly updated during a Save operation (#106)
Fixes #105 closes #105

This is more serious that the bug above appears to be.

When calling `CopyFrom` on `Gateways`, and in some cases on `Scenes`,
the target Gateway or Scene would end up pointing to the wrong house
(the "from" one as opposed to the target one) after the copy. This would
potentially have a bunch of consequences, including sending
notifications used to animate the "Sync" icon to the wrong house, and
preventing the "from" house to be garbage collected.
 
Changed implementation of `Gateways.CopyFrom` to ensure that the
gateways are pointing to the right house model after the copy. This
involved implementing `Gateway.CopyFrom`. Since I was there, I also
simplified the `Gateway` constructors.

Fixed a bug in `Scenes.CopyFrom` that would also make the scene points
to the wrong scene list and house model in certain cases.

Also added immutable Id to House to help with identifying the house
while debugging.