Double checked CopyFrom implementations and added one more merge test (#107)
The CopyFrom methods are used to apply the model obtained by merging
remote and local changes back onto the model connected to the UI to get
the UI to update. They should make a deep copy of the model objects
(House, Gateways, Devices, AllLinkDatabases, Scenes, SceneMembers, while
maintaining the integrity of the model tree. Some objects hold a
reference back to their immediate container or to the House, and this
should not be changed, while the rest of the state should be copied.

A review of the implementations of CopyFrom identified a few bugs where
either 1/ a reference back up toward the House was incorrectly modified
or 2/ not all state was copied. Fixed these issues.

I also added a new unit-test adding a device locally while removing one
remotely. Will add more scenarios shortly.