Christian/new unit tests (#110)
- Added more unit-tests testing merging local changes with remote
changes (changes to the persisted file from another instance of the app,
usually on a different device) and persisting to file.

- Fixed two bugs having to do with attempting to create duplicate
devices or duplicate links in that process.
- Added a unit-test removing a scene locally and remotely. 
- Added unit-test testing scene creation resulting into a conflict of
scene Id. Created `ModelRecorderContext` to perform scene id adjustment
when a new id has to be assigned to a scene because of conflict.