Ensured that new devices or new scenes are shown and in view (#118)
This change ensures that when adding a device or a scene, its view model
will be added to the presented list of view models and selected and
brought into view, regardless of the filtering on that list.
This change also simplifies the `NewSceneDialog`, which now just return
a scene name. It's now up to the caller to add the new scene to the
model and the view model.