Handle last known to be disconnected device during sync (#120)
Now handling the case when during sync a device is last known not to be
connected by pinging the device and requesting a new sync pass if the
ping succeeds. Otherwise, device marked disconnected for whatever reason
will never have a chance to sync again, until directly viewed by the
user (which will ping a potentially reconnect them).

Also simplified DeviceNeedsSync and renamed to NeedsSync since it's a
method of Device.