Now scrolling to the added scene member when adding a new one (#121)
This is made a bit more difficult than meets the eyes because the
`GridView` showing the scene members is not actually scrolling. Instead,
it is fully expanded inside a `ScrollViewer` that also contains other
UX. In addition, the `ScrollViewer` and `GridViews` are both in
templated content and are not accessible with the regular FindName
method.

To achieve this:
- Created a few utility helpers to operate on the XAML visual tree that
allows 1/ to find elements by name or type, going through control and
template boundaries, and 2/ to scroll an item of a `GridView` in view
using a different `ScrollViewer` than the one of the `GridView`.
- Combined these helpers with the functionality previously in XAMLUtils
to reduce duplication and updated code that depended on it.
- Used them to scroll a scene member into view that the user is adding.
In the Scene view, a `ScrollViewer` contains the `GridView` of scene
members and more UX. The `GridView` itself is not scrolling.
- Future PR will do the same for database records (links). 
- Added `Equals` method to `MemberViewModel` to search by value in the
scene member lists instead of by reference. Contrary to other view
models, Scene Members have no unique key and no guaranty that they don't
get duplicated. So searching by reference is meaningless and was
probably the source of bugs.
- Made the `UIScheduler` public to use in application layer. Used it to
delay the scrolling into view a bit to let the `GridView` the
opportunity to update with the new scene member before attempting to
scroll to it.

Fixes #68 Closes #68