Fixed offset computation in `ScrollItemIntoView` (#122)
`FrameworkElement.TransformToVisual` returns the coordinate transform
between two framework elements. It is used here to compute the transform
between a Framework element and a `ScrollViewer` in its ancestor chain.
In that case `TransformToVisual` returns the transform with respect to
the viewport of the `ScrollViewer`, not its content. This would make my
implementation of `ScrollItemIntoView` not bring the element in the
right place whenever the `ScrollViewer` was scrolled.

The fix is to compute the transform with respect to the content of
`ScrollViewer` instead.

This change also renames a few variables for consistency and
readability.