Workaround to make the `LoadControl` at top of Device view work on Uno.  (#125)
Issue #124

It appears that on Uno, binding the events of a content control in a
subtree with the `x:Load` attribute does not work. The event handlers
are null despite the binding. This all works just fine on WindowsAppSdk
(WinUI3).

Replacing `x:Load` by `Visibility` appears to fix the issue on
WindowsAPpSdk and Uno based platforms, so going with that for now, and I
will create a simple repro to investigate further.