Better fix to make the `LoadControl` in Device view work on Uno. (#126)
As stated in PR #125, binding to an event in the view model does not
work on Uno if there is an x:Load in the ancestor chain (it does work in
WinUI3). Since whether binding to an event in the view model should work
or not seems unclear, I changed the events to commands, which appears to
be the recommended way to handle this.

I changed `LoadFullOn`, `LoadOn`, `LoadOff` events to commands instead
and adapted the dependent code accordingly.

Since commands can easily pass a parameter, I removed the
`LoadfullOnCommand` and instead we use `LoadOnCommand` with a parameter
indicating the level, which should be 1.0d for full on.