After user adds a new link, bring it into view. (#130)
This change uses the helpers in `XAMLHelpers` to bring a newly added
link into view in the Device view. It also ensures that the link is
added to the `LinkListViewModel` regardless of current filter, giving
feedback to the user that the link has indeed been added.

We now `Add/RemoveObserver` of the `AllLinkDatabase` in
`ActiveStateChange` instead of at creation of the `LinkListViewModel`
which avoid adding the observer several times, which then would make us
add a new `LinkViewModel` several times.

Also added Equals method to `LinkViewModel` to look them up by value in
`ListLinkViewModel` for example using `IndexOf`. This is important as
the `LinkViewModel` that we have in hand might not always be the same
instance as what is in the `LinkListViewModel`, forcing us to look for
the one in the list by value.