UX and code improvements and simplifications to link grid views (#133)
- Removed the `LinkListViewModel` change tracking logic and instead now
rebuild the lists each time the device or channel view gets activated.
There was no major benefit to tracking the changes, since they happen
relatively rarely.
- Now tracking the selection on link lists and restoring it after
rebuilding the lists. This ensure a consistent preservation of the
selected link in each list (as long as the same `LinkHostViewModel` is
used).
- We now have a single selected link across controller and responder
links, simplifying the UX.
- Ensured that the HubChannel view always show the link group headers
with the "Add" button, even when that group is empty (no link in it)
- Renamed `GetLinkIndex` to `GetLinkViewModeIndex` for consistency with
`GetLinkViewModel`.
- Removed now unnecessary code in `LinkListViewModel`.