Fixed exception at launch on DESKTOP framework (#134)
When launching the app on the Desktop Uno framework, a exception was
raised early on because of an initialized variable in Uno's
`DispatcherQueue.GetForCurrentThread()`. This was because of missing
code in the Main function in `desktop\program.cs`, made necessary after
the upgrade to Uno 6. I imported this code from a new sample app, which
fixed the issue.