Upgraded to .net9.0 and Uno 6.0.146 (#137)
Upgraded to .net9.0 and Uno 6.0.146.
Removed `net9.0-desktop` and `net9.0-browserwasm` from the list of
frameworks for the libraries and replaced by `net9.0` based on comments
[here](https://github.com/unoplatform/uno/issues/20312#issuecomment-2887282787).
Removed dependencies on `UniversalImageLoader` which were causing
problems and are not necessary anymore.
Added code in Main function in program.cs for platforms Desktop and
WebAssembly.
Adapted README.cs as needed.