Prepare for turning on SkiaRenderer but did not turn it on yet. (#138)
When turning on the `SkiaRenderer` feature, a couple of behavior
breaking changes needed to be handled. In particular:
- The `Main` function for Desktop needs to initialize some additional
services, or we get a crash at startup.
- A workaround for a semantic difference between Uno and WinUI3 in the
parameters of `DataTemplateSelector.SelectTemplateCore` needs to be
extended to all platforms using Uno, not just Desktop and WASM,
- Uno bug 20601 in WithUnoHelpers requires a workaround for Android. 

With these changes, Skia Renderer can be turned on by adding
`SkiaRenderer` to `UnoFeatures` in `UnoApp.csproj`:
```
    <UnoFeatures>
      Lottie;
      ...
      Toolkit;
      SkiaRenderer;
    </UnoFeatures>
```

This change does not turn on `SkiaRenderer` yet. I want to run it
locally for a while to see if there are more changes like these to
address before turning it on by default.