Fixed issue preventing to run as an unpackaged app. (#152)
Unpackaged WindowsAppSDK apps do not have access to
`ApplicationDataContainer` which was used to store app settings. This
would generate a system exception. Implemented an alternative way of
storing settings, in a file in the app data, for unpackaged apps.

This now makes it possible to build and run the solution as an
unpackaged app on WindowsAppSDK. Modified README.md to reflect that.