Derive version numbers from Git repo using MinVer (#175)
With this change:
- We use MinVer to compute the version numbers based on certain tags in
the Git repo. To control the version number, tags have to match to the
following format:
```
v<Major>.<Minor>.<Patch>[-beta|preview]
```
where <Major>, <Minor> and <Patch> are decimal numbers

Examples: 
```
v0.9.4-beta
v1.2.7
v3.5.6
```
A package built for the MS Store cannot contain `-beta` but has a 4th
number, always set to '0'.
Tags can be on any branch.
The closest version format matching tag to the top of the branch is
used.
A build error will happen if no tag is found. 
Pushing the tag to the remote branch builds and publishes to a new
GitHub release using `release.yml`. The builds outputs are attached to
the release.