Prevent Package.appxmanifest from incorrectly changing at random times.  (#184)
Issue: In `Package.appxmanifest` `Identity` is being changed at
seemingly random times by the `XmlPokes` in `UnoApp.csproj` with the
wrong information (e.g., version number set to '...0').

It appears to be a result of the Design Time builds that VS performs for
syntax coloring, IntelliSense and other purpose. For efficiency, these
builds do not run certain targets such as the MinVer one, and the
version is never acquired. These builds also do not perform any
packaging so it's not actually necessary to obtain the package identity
version.

The fix is to disable target `UpdateAppxManifest` in Design Time builds,
which normally would trigger the version acquisition using MinVer. That
ensures that we do not touch the `Package.appxmanifest` for Design Time
builds.

Also simplified target dependencies for easier understanding and cleaned
up logging.

Checked-in default state for `Package.appxmanifest` with 0,0,0,0 as
version number.

Upgraded to MinVer 6.0.0.