Clear Gateway HttpClient cache when copying from another gateway (#194)
Fixes #192. Gateway objects contain an `HttpClient` instance that is
obtained once and get reused until cleared. When copying values from
another Gateway instance in `Gateway.CopyFrom`, we were not clearing the
`HttpClient` even though the new property values might require obtaining
a new instance. This would happen in particular after the user entered
wrong Hub settings, which would create an incorrect `HttpClient`
resulting in the devices showing disconnected. When the user then
corrected these values, the `HttpClient` would not get reacquired and
the devices would not reconnect. The fix is simply to null out
`httpClient` in `CopyFrom` to force reacquiring it next time
`HttpClient.Get` is called.

This change also edits a few comments to stress that calling `CopyFrom`
is necessary (as opposed to creating new instances) in order to generate
property change notifications and deltas.