Display OnLevel as % in the GUI and Console (#200)
Now displaying OnLevel consistently in % everywhere in the UI and in
the console. In addition, in the console, the user can type a percentage
for OnLevel instead of the underlying byte value. The underlying byte
value is still being displayed an accepted as input in the console for
reference. We might decide to remove it altogether later based on
experience.