Fixed incorrect UI Settings state shown after scanning network for hub. (#204)
The Hub Settings info (ip address, device Id) was incorrectly captured
before the call to `TryPushNewGatewayAsync` could retrieve it. Also, the
discovery state of the Hub (found or not) would be overwritten while
setting `HubIpAddress`, showing the Hub as not found, even when it was
actually found.

- Ensures that `TryPushNewGatewayAsync` has run successfully before
capturing Hub ip address and Insteon id for the Settings view.
- Ensures that setting the hub discovery state to "Found" is done last
and not overridden.
- Removed unnecessary shortcut to `Holder.House.Gateway.DeviceId`.
- Add some comments to facilitate future understanding.
- Rename local var hubIpAddress to hubIp to avoid confusion with private
member hubIPAddress (different in caps only)