Ensure that we pick-up latest Settings changes when scanning the network for the hub (#205)
`_verificationClient` was cached and only created once for performance.
This is wrong because we need to pick up the latest changes from the
user in the Hub settings information such as username and password for
example, and we were not. We could recreate and cache the verification
client at the start of a scan but given that this is only used on
devices with a specific port open, it's unlikely to be used more than
once or twice per scan. So, we just recreate `verificationClient` each
time we need it to ensure that we pick up the user latest changes in the
Hub Settings.