Ensured StatusTextLogger could be called from non-UI thread (#206)
`StatusTextLogger` could only be called from the main (UI) thread for a
couple of reason:
1. `messageTimer` was a `DispatchTimer` which only works on the UI
thread.
2. the callback to display text in the status bar needed to be marshaled
to the UI thread.

These issues could cause a COM exception inside the timer code in some
conditions and could in particular make the scan of the local network
(to find the Hub) fail. This change removes these limitations:
- Used a System Timer rather than a `DispatchTimer`, which is UI thread
only.
- Marshall to main thread before invoking the callback.
- Remove redundant logging and modified logged text for clarity.